requirements.txt
filelock==3.13.1
micropython-phew==0.0.3
pipkin==2.0b2
pyserial==3.5
